fix build on Windows

+++ include/uni_can_devinfo.h
@@ -12,7 +12,6 @@ extern "C" {
 #include <stdint.h>
 
 
-
 //
 // Defines
 //
@@ -20,19 +19,17 @@ extern "C" {
 #define UNI_CAN_DEVINFO_FIELDLEN (32U)
 
 
-
 //
 // Typedefs
 //
 
 typedef struct {
-
-  char device_manufacturer [UNI_CAN_DEVINFO_FIELDLEN];
-  char device_model        [UNI_CAN_DEVINFO_FIELDLEN];
-  char device_sn           [UNI_CAN_DEVINFO_FIELDLEN];
-  char device_type         [UNI_CAN_DEVINFO_FIELDLEN];
-  uint32_t device_index;
-  uint32_t device_chancnt;
+    char device_manufacturer[UNI_CAN_DEVINFO_FIELDLEN];
+    char device_model[UNI_CAN_DEVINFO_FIELDLEN];
+    char device_sn[UNI_CAN_DEVINFO_FIELDLEN];
+    char device_type[UNI_CAN_DEVINFO_FIELDLEN];
+    size_t device_index;
+    uint32_t device_chancnt;
 } uni_can_devinfo_t;
